remove urldialog, replace with query toggle state hook
