+++ src/browser/query/index.ts
@@ -27,3 +27,5 @@ export function setQueryParam({ key, router, push, value }: SetQueryParamOptions
     router.replace({ query: router.query }, undefined, { shallow: true }).catch(console.error);
   }
 }
+
+
